requirements.txt
pyrogram==2.0.106
tgcrypto
motor
pymongo
aiohttp
python-dotenv
dnspython
